refactor and move tests to spec files
