+++ dpp/src/document/mod.rs
@@ -21,7 +21,7 @@ pub mod document_factory;
 pub mod document_validator;
 pub mod errors;
 pub mod generate_document_id;
-mod state_transition;
+pub mod state_transition;
 
 const IDENTIFIER_FIELDS: [&str; 3] = ["$id", "$dataContractId", "$ownerId"];
