+++ dpp/src/document/state_transition/documents_batch_transition/mod.rs
@@ -1,4 +1,3 @@
-use log::trace;
 use std::collections::HashMap;
 use std::convert::TryInto;
