+++ dpp/src/document/state_transition/documents_batch_transition/validation/state/fetch_documents.rs
