+++ dpp/src/document/state_transition/documents_batch_transition/validation/state/mod.rs
@@ -1,4 +1,4 @@
 pub mod execute_data_triggers;
-pub mod fetch_documents_factory;
+pub mod fetch_documents;
 pub mod validate_documents_batch_transition_state;
 pub mod validate_documents_uniqueness_by_indices;
