+++ dpp/src/tests/document/mod.rs
@@ -1,0 +1,1 @@
+pub mod state_transitions;
