+++ dpp/src/tests/document/state_transitions/mod.rs
@@ -1,0 +1,1 @@
+mod validation;
