+++ dpp/src/tests/document/state_transitions/validation/mod.rs
@@ -1,0 +1,3 @@
+mod validate_documents_batch_transition_state_spec;
+mod validate_documents_uniqueness_by_indices_spec;
+mod validate_partial_compound_indices_spec;
