+++ dpp/src/tests/mod.rs
@@ -4,5 +4,6 @@ mod identifier_spec;
 pub mod utils;
 
 mod data_contract;
+mod document;
 mod identity;
 mod version;
